refactor NSTemplateSet/Namespace ownershipA
add a finalizer on the NSTemplateSet resources
so that the controller is called when an NSTemplateSet
is deleted. When this happens, the controller sets the
status of the NSTemplateSet resource to `ready=false/reason=terminating`
and deletes the "child" namespaces one by one.
Once all namespaces have been deleted (or marked for deletion, ie
they have a deletion timestamp), the controller removes the
finalizer on the NSTemplateSet resource so that the GC
can delete it for good.

also:
- when creating the NSTemplateSet in the UserAccount controller,
add the finalizer aforementioned
- do not set the ownership to the NSTemplateSet on the user namespaces
- use a custom Event Handler to map the event on a namespace to its
associated NSTemplateSet (if a specific label is found)
- add permission to the member-operator service account
to delete namespaces

Fixes https://issues.redhat.com/browse/CRT-429

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>