disable CR metrics
by default, this function wants to use all out CRDs, but when we have a "host-only"
cluster, this metrics server fails to load the CRDs from etcd, which causes CrashLoops.
If we really need this metrics server, then we should taylor the list of CRDs to expose.

Keeping the `serveCRMetrics` for now, but with a comment to avoid linter issue,
since it's not used.

Also, address pending lint issues with fixes or comments

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>