refactor: add MasterUserRecord.Spec.TierName field (only)
Just a module update with the new field in `MasterUserRecord.Spec`,
but no change in the business logic.
Applying the same module update on both host-operator and member-operator
so that when the business logic is applied, both operators can use the
new field.

This is a workaround the limitations of e2e test pairing.
See [discussion](https://github.com/codeready-toolchain/host-operator/pull/539#issuecomment-975630203)

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>