fix(#343): Add support for scale subresource
Update the idler component to also support deployments managed by scale subresource. In case a deployment is owned by a custom resource that holds a scale subresource the idler will patch the scale resource instead of the deployment directly.

This makes sure that the idler is able to terminate deployments owned by scale resources, too.

This is being used by Camel K workloads such as Integration and KameletBinding custom resources.