fix(#343): Support scale resources that do not have specReplicasPath field set
The replicas field defined in specReplicasPath may be optional in the custom resource. In this case getting the scale resource fails with internal server error. In this case try to set a new scale resource with replicas=0. This fixes idling of Camel K integrations and bindings that do not explicitly have a `spec.replicas` field set and default to replicas=1.

Signed-off-by: Christoph Deppisch <cdeppisch@redhat.com>