ci: bump actions to silence node12 warnings (#481)
As of [June 2023][1], GitHub no longer uses node v12 within its actions
runner by default.  They have been encouraging action authors to switch
to node v16 and [encouraging users][2] to use the new versions of the
actions when they become available, with the intent on removing node v12
sometime in the near future.

Bump the versions of the affected actions to bring this in line.

[1]: https://github.blog/changelog/2023-06-13-github-actions-all-actions-will-run-on-node16-instead-of-node12-by-default/
[2]: https://github.com/codeready-toolchain/host-operator/actions/runs/6421632076

Signed-off-by: Andy Sadler <ansadler@redhat.com>
Co-authored-by: Alexey Kazakov <alkazako@redhat.com>